refactor(mcp): extract port helpers into mcp/ports.py
Move MCP_HOST, _read_env_port, _port_available, and _select_port from
server.py into a new mcp/ports.py module so CLI commands (doctor, mcp
serve) can read the env var or probe a port without importing the
FastAPI/uvicorn server stack.

- doctor_cmd and mcp_cmd now import from mcp.ports directly
- server.py imports the symbols back from ports for its own use
- _serve_in_thread return type narrowed from Any to threading.Thread
- Registry.resolve_tool_names: introduce _mutating_allowlist_for(role)
  to collapse the duplicated reviewer/non-reviewer comprehensions (D9
  dual allowlist stays intact)
